# Publish - avoid async state machine when possible