@mtmk 2024-02-04
Semaphore here is giving us better performance clearly. Main gain is probably not having to manage timeouts with cancellation tokens.