@caleblloyd 2024-02-04
Could also be seeing gains from skipping the async state machinery